requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
python-dotenv
docxtpl
pymysql
cryptography
flask-socketio
eventlet